# openapi.yaml
openapi: 3.0.3
info:
  title: Kinloop API
  version: 1.0.0
  description: OpenAPI specification for Kinloop backend (generated from controllers and DTOs)
servers:
  - url: https://kinloop-be.onrender.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ChildId:
      name: childId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    SessionStatus:
      type: string
      enum: [IN_PROGRESS, COMPLETED, ABANDONED]
    AgeBand:
      type: string
      enum: [BAND_0_12, BAND_12_24, BAND_24_48, BAND_48_72]
    Gender:
      type: string
      enum: [FEMALE, MALE, OTHER, UNDISCLOSED]
    TriggerReason:
      type: string
      enum: [INITIAL, AGE_BAND, MANUAL]
...
    CreateChildRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 255
        birthDate:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/Gender'
      required: [birthDate]
    UpdateChildRequest:
      allOf:
        - $ref: '#/components/schemas/CreateChildRequest'
    UpdateChildResponse:
      type: object
      required: [childId, displayName, birthDate, ageMonths, ageBand, questionnaireRestarted]
      properties:
        childId:
          type: integer
          format: int64
        fullName:
          type: string
        displayName:
          type: string
        birthDate:
          type: string
          format: date
        ageMonths:
          type: integer
        ageBand:
          $ref: '#/components/schemas/AgeBand'
        gender:
          $ref: '#/components/schemas/Gender'
        questionnaireRestarted:
          type: boolean
          description: True when a corrected birth date changes the age band and a new questionnaire is required.
...
    SessionSummaryResponse:
      type: object
      properties:
        sessionId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/SessionStatus'
        triggerReason:
          $ref: '#/components/schemas/TriggerReason'
        totalQuestions:
          type: integer
        answeredCount:
          type: integer
        nextQuestionCode:
          type: string
    CreateChildResponse:
      type: object
      properties:
        childId:
          type: integer
          format: int64
        fullName:
          type: string
        gender:
          $ref: '#/components/schemas/Gender'
        displayName:
          type: string
        birthDate:
          type: string
          format: date
        ageMonths:
          type: integer
        ageBand:
          $ref: '#/components/schemas/AgeBand'
        session:
          $ref: '#/components/schemas/SessionSummaryResponse'
...
paths:
...
  /api/children:
    post:
      summary: Create a child for current parent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChildRequest'
      responses:
        '201':
          description: Child created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateChildResponse'
  /api/children/{childId}:
    put:
      summary: Update an owned child's identity details from onboarding or profile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChildId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChildRequest'
      responses:
        '200':
          description: Child updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateChildResponse'
        '404':
          description: Child does not belong to the authenticated parent
...
security:
  - bearerAuth: []